Fixed async bug in createNewUser, added toasts for LogInAPI
