+++ app/src/main/java/softeng/eventplanning/MainActivity.java
@@ -47,7 +47,7 @@ public class MainActivity extends AppCompatActivity {
                 final LogInAPI asyncT = createAsyncTask();
                 asyncT.setsomething(create_user_info);
                 if(asyncT.getStatus() != AsyncTask.Status.RUNNING) {
-                    Log.d("CALLED", "yup");
+                    //Log.d("CALLED", "yup");
                     String user_password = password.getText().toString();
                     String username = username_input.getText().toString();
                     create_user_info[0] = username;
@@ -63,20 +63,10 @@ public class MainActivity extends AppCompatActivity {
         getSupportActionBar().hide();
 
     }
-    public LogInAPI createAsyncTask(){
+    private LogInAPI createAsyncTask(){
         LogInAPI api = new LogInAPI();
         api.signupActivity(this);
         return api;
-    }
-
-    public void eventClicked(View view){
-        // TODO open event's page when clicked
-    }
-
-    public void settingsClicked(View view){
-        // TODO open settings view when clicked
-
-
     }
     public void createEventClicked(MenuItem item){
         Intent createEventChange = new Intent(this,createEvent.class);
@@ -87,64 +77,22 @@ public class MainActivity extends AppCompatActivity {
         Intent intent = new Intent(this, MainActivity.class);
         startActivity(intent);
     }
-
-    public void setLocation(View v){
-        Intent eventLocationChange = new Intent(this, createEventLocation.class);
-        startActivity(eventLocationChange);
-    }
-
-    public void setDate(View v){
-        Intent eventDateChange = new Intent(this, createEventDate.class);
-        startActivity(eventDateChange);
-    }
-
-    public void setTime(View v){
-        Intent eventTimeChange = new Intent(this, createEventTime.class);
-        startActivity(eventTimeChange);
-    }
-
-    public void back(View v){
-        Intent createEventChange = new Intent(this,createEvent.class);
-        startActivity(createEventChange);
-    }
-
-
-    public void sendLogin(View v) {
-        //public final static String EXTRA_MESSAGE = "softeng.eventplanning.MESSAGE";
-        /*
-        Intent intent = new Intent(this, sendLogin.class);
-        EditText getUsername = (EditText) findViewById(R.id.edit_username);
-        EditText getPassword = (EditText) findViewById(R.id.edit_password);
-        String username = getUsername.getText().toString();
-        String password = getPassword.getText().toString();
-        //intent.putExtra(EXTRA_MESSAGE, username,password);
-        intent.putExtra(username,password);
-        startActivity(intent);
-        */
-        // TODO add try catch if user login was successful to call below
-
-        setupTabsView(v);
-    }
-
     public void setupTabsView(View v) {
         getSupportActionBar().show();
         Intent tabViewChange = new Intent(this, tabView.class);
         startActivity(tabViewChange);
-
     }
-
     public void createAccount(View v){
         Intent intent = new Intent(this, CreateNewUser.class);
         startActivity(intent);
     }
-
+    //Menu
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
         getMenuInflater().inflate(R.menu.settings_menu, menu);
         return true;
     }
-    //Menu
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
         // Handle action bar item clicks here. The action bar will
